Create tests for async task
